Flatten the `js/src/components/free-listings/configure-product-listings/shipping-time` directory structure.
