+++ js/src/components/free-listings/configure-product-listings/shipping-time-section.js
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import Section from '~/wcdl/section';
 import AppDocumentationLink from '~/components/app-documentation-link';
-import ShippingTimeSetup from './shipping-time/shipping-time-setup';
+import ShippingTimeSetup from './shipping-time-setup';
 
 /**
  * @fires gla_documentation_link_click with `{ context: 'setup-mc-shipping', link_id: 'shipping-read-more', href: 'https://support.google.com/merchants/answer/7050921' }`
