+++ js/src/components/free-listings/configure-product-listings/shipping-time-setup/add-time-button/add-time-modal.js
@@ -14,52 +14,30 @@ import AppModal from '~/components/app-modal';
 import VerticalGapLayout from '~/components/vertical-gap-layout';
 import SupportedCountrySelect from '~/components/supported-country-select';
 import validateShippingTimeGroup from '~/utils/validateShippingTimeGroup';
-import MinMaxShippingTimes from '../../../min-max-shipping-times';
-import '../index.scss';
+import MinMaxShippingTimes from '../min-max-shipping-times';
 
 /**
- *Form to edit time for selected country(-ies).
+ * Form to add a new time for selected country(-ies).
  *
  * @param {Object} props
- * @param {Array<CountryCode>} props.audienceCountries List of all audience countries.
- * @param {AggregatedShippingTime} props.time
- * @param {(newTime: AggregatedShippingTime, deletedCountries: Array<CountryCode>) => void} props.onSubmit Called once the time is submitted.
- * @param {(deletedCountries: Array<CountryCode>) => void} props.onDelete Called with list of countries once Delete was requested.
- * @param {Function} props.onRequestClose Called when the form is requested ot be closed.
+ * @param {Array<CountryCode>} props.countries A list of country codes to choose from.
+ * @param {Function} props.onRequestClose
+ * @param {function(AggregatedShippingTime): void} props.onSubmit Called with submitted value.
  */
-const EditTimeModal = ( {
-	audienceCountries,
-	time,
-	onDelete,
-	onSubmit,
-	onRequestClose,
-} ) => {
+const AddTimeModal = ( { countries, onRequestClose, onSubmit } ) => {
 	const [ dropdownVisible, setDropdownVisible ] = useState( false );
 
-	// We actually may have times for more countries than the audience ones.
-	const availableCountries = Array.from(
-		new Set( [ ...time.countries, ...audienceCountries ] )
-	);
-
-	const handleDeleteClick = () => {
-		onDelete( time.countries );
-	};
-
 	const handleSubmitCallback = ( values ) => {
-		const remainingCountries = new Set( values.countries );
-		const removedCountries = time.countries.filter(
-			( el ) => ! remainingCountries.has( el )
-		);
-
-		onSubmit( values, removedCountries );
+		onSubmit( values );
+		onRequestClose();
 	};
 
 	return (
 		<Form
 			initialValues={ {
-				countries: time.countries,
-				time: time.time,
-				maxTime: time.maxTime,
+				countries,
+				time: 0,
+				maxTime: 0,
 			} }
 			validate={ validateShippingTimeGroup }
 			onSubmit={ handleSubmitCallback }
@@ -83,21 +61,13 @@ const EditTimeModal = ( {
 						) }
 						buttons={ [
 							<AppButton
-								key="delete"
-								isTertiary
-								isDestructive
-								onClick={ handleDeleteClick }
-							>
-								{ __( 'Delete', 'google-listings-and-ads' ) }
-							</AppButton>,
-							<AppButton
 								key="save"
 								isPrimary
 								disabled={ ! isValidForm }
 								onClick={ handleSubmit }
 							>
 								{ __(
-									'Update shipping time',
+									'Add shipping time',
 									'google-listings-and-ads'
 								) }
 							</AppButton>,
@@ -110,20 +80,18 @@ const EditTimeModal = ( {
 									'If customer is in',
 									'google-listings-and-ads'
 								) }
-								countryCodes={ availableCountries }
+								countryCodes={ countries }
 								onDropdownVisibilityChange={
 									setDropdownVisible
 								}
 								{ ...getInputProps( 'countries' ) }
 							/>
-
 							<div className="label">
 								{ __(
-									'Then the estimated shipping times displayed in the product listing are',
+									'Then the estimated shipping times displayed in the product listing are:',
 									'google-listings-and-ads'
 								) }
 							</div>
-
 							<Flex
 								direction="column"
 								className="gla-countries-time-input-container"
@@ -137,6 +105,7 @@ const EditTimeModal = ( {
 										handleBlur={ handleIncrement }
 										handleIncrement={ handleIncrement }
 									/>
+
 									<ul className="gla-validation-errors">
 										<li>{ errors.time }</li>
 									</ul>
@@ -150,7 +119,7 @@ const EditTimeModal = ( {
 	);
 };
 
-export default EditTimeModal;
+export default AddTimeModal;
 
 /**
  * @typedef { import("~/data/actions").AggregatedShippingTime } AggregatedShippingTime
