+++ js/src/components/free-listings/configure-product-listings/shipping-time-setup/add-time-button/add-time-modal.test.js
@@ -7,7 +7,7 @@ import { render, fireEvent, waitFor } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import AddTimeModal from './';
+import AddTimeModal from './add-time-modal';
 import useMCCountryTreeOptions from '~/components/supported-country-select/useMCCountryTreeOptions';
 
 jest.mock( '~/components/supported-country-select/useMCCountryTreeOptions' );
