+++ js/src/components/free-listings/configure-product-listings/shipping-time-setup/countries-time-input/edit-time-modal.test.js
@@ -7,7 +7,7 @@ import { render, fireEvent, waitFor } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import EditTimeModal from './';
+import EditTimeModal from './edit-time-modal';
 import useMCCountryTreeOptions from '~/components/supported-country-select/useMCCountryTreeOptions';
 
 jest.mock( '~/components/supported-country-select/useMCCountryTreeOptions' );
